+++ orix/plot/__init__.py
@@ -16,7 +16,6 @@
 # You should have received a copy of the GNU General Public License
 # along with orix. If not, see <http://www.gnu.org/licenses/>.
 #
-
 """Extensions of Matplotlib's projections framework for plotting
 :class:`~orix.vector.Vector3d`, :class:`~orix.quaternion.Rotation`,
 :class:`~orix.quaternion.Orientation`,
@@ -25,6 +24,13 @@
 """
 
 import lazy_loader
+
+from orix.plot._plot import register_projections
+
+# TODO: Remove this call once relying on the old side-effect of
+# importing orix.plot to register these projections has little-to-no
+# consequence
+register_projections()
 
 # Imports from stub file (see contributor guide for details)
 __getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)